create a new file nodes.go and move node part codes from cluster.go into nodes.go
